// apps/web/src/shared/api/generated/model/notesControllerListBookArchiveParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */
import type { NotesControllerListBookArchiveCategoryItem } from "./notesControllerListBookArchiveCategoryItem";
import type { NotesControllerListBookArchiveFilter } from "./notesControllerListBookArchiveFilter";
import type { NotesControllerListBookArchiveHasChapter } from "./notesControllerListBookArchiveHasChapter";
import type { NotesControllerListBookArchiveHasPage } from "./notesControllerListBookArchiveHasPage";
import type { NotesControllerListBookArchiveSort } from "./notesControllerListBookArchiveSort";

export type NotesControllerListBookArchiveParams = {
  /**
   * @maxItems 100
   */
  category?: NotesControllerListBookArchiveCategoryItem[];
  /**
   * @maxItems 100
   * @items.minLength 1
   * @items.maxLength 100
   */
  customCategory?: string[];
  /**
   * @maxItems 100
   * @items.pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  author?: string[];
  /**
   * @maxItems 100
   * @items.pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  book?: string[];
  hasChapter?: NotesControllerListBookArchiveHasChapter;
  hasPage?: NotesControllerListBookArchiveHasPage;
  /**
   * @maxLength 100
   */
  search?: string;
  filter?: NotesControllerListBookArchiveFilter;
  /**
   * @minimum 1
   * @maximum 21474836
   */
  pageNumber?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
  sort?: NotesControllerListBookArchiveSort;
};