// apps/web/src/shared/api/generated/model/notesControllerListSeriesArchiveParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */
import type { NotesControllerListSeriesArchiveCategoryItem } from "./notesControllerListSeriesArchiveCategoryItem";
import type { NotesControllerListSeriesArchiveFilter } from "./notesControllerListSeriesArchiveFilter";
import type { NotesControllerListSeriesArchiveReadingItem } from "./notesControllerListSeriesArchiveReadingItem";
import type { NotesControllerListSeriesArchiveSort } from "./notesControllerListSeriesArchiveSort";
import type { NotesControllerListSeriesArchiveStatusItem } from "./notesControllerListSeriesArchiveStatusItem";

export type NotesControllerListSeriesArchiveParams = {
  /**
   * @maxItems 100
   */
  category?: NotesControllerListSeriesArchiveCategoryItem[];
  /**
   * @maxItems 100
   * @items.minLength 1
   * @items.maxLength 100
   */
  customCategory?: string[];
  /**
   * @maxItems 100
   * @items.pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  author?: string[];
  /**
   * @maxItems 100
   * @items.minLength 1
   * @items.maxLength 64
   */
  genre?: string[];
  /**
   * @maxItems 100
   */
  reading?: NotesControllerListSeriesArchiveReadingItem[];
  /**
   * @maxLength 100
   */
  search?: string;
  /**
   * @maxItems 100
   * @items.pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  series?: string[];
  /**
   * @maxItems 100
   */
  status?: NotesControllerListSeriesArchiveStatusItem[];
  filter?: NotesControllerListSeriesArchiveFilter;
  /**
   * @minimum 1
   * @maximum 21474836
   */
  pageNumber?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
  sort?: NotesControllerListSeriesArchiveSort;
};